+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.3.0
 
 * Reworked instrumentation(breaking if used)
-* AtomicIsizes in inner_pool to allow "temporary underflow)
+* AtomicIsizes in inner_pool to allow "temporary underflow"
 * Config from environment
 
 ## 0.2.0